requirements.txt
yfinance 
scikit-learn
tensorflow-macos
pandas
tensorflow-metal